-count 분리 -user가 달에 처음 올린 고지서에만 카운트가 올라가게 변경(사용기간 카운트)

+++ server/src/main/java/com/oss/carbonadministrator/service/image/ElectricityImageService.java
@@ -31,10 +31,6 @@ public class ElectricityImageService {
             throw new HasNoUserException("해당하는 유저가 존재하지 않습니다.");
         }
 
-        User targetUser = user.get();
-        targetUser.setCount(targetUser.getCount() + 1);
-        userRepository.saveAndFlush(targetUser);
-
         Optional<Bill> targetBill = billRepository.findBillByEmailAndYearAndMonth(user.get().getEmail(), year, month);
 
         if (targetBill.isEmpty()){
@@ -49,6 +45,16 @@ public class ElectricityImageService {
         }
 
         Bill bill = targetBill.get();
+
+        if(bill.getElectricityInfoList() == null){
+            User targetUser = user.get();
+            targetUser.setElecCount(targetUser.getElecCount() + 1);
+            userRepository.saveAndFlush(targetUser);
+
+            bill.setElectricityInfoList(recognizedElecData);
+
+            return billRepository.saveAndFlush(bill);
+        }
         electricityRepository.delete(bill.getElectricityInfoList());
         bill.setElectricityInfoList(recognizedElecData);
